+++ compiled/sharp/src/operations.cc
@@ -267,23 +267,6 @@ namespace sharp {
     }
   }
 
-  /*
-    Insert a tile cache to prevent over-computation of any previous operations in the pipeline
-  */
-  VImage TileCache(VImage image, double const factor) {
-    int tile_width;
-    int tile_height;
-    int scanline_count;
-    vips_get_tile_size(image.get_image(), &tile_width, &tile_height, &scanline_count);
-    double const need_lines = 1.2 * scanline_count / factor;
-    return image.tilecache(VImage::option()
-      ->set("tile_width", image.width())
-      ->set("tile_height", 10)
-      ->set("max_tiles", static_cast<int>(round(1.0 + need_lines / 10.0)))
-      ->set("access", VIPS_ACCESS_SEQUENTIAL)
-      ->set("threaded", TRUE));
-  }
-
   VImage Threshold(VImage image, double const threshold, bool const thresholdGrayscale) {
     if (!thresholdGrayscale) {
       return image >= threshold;
@@ -358,4 +341,18 @@ namespace sharp {
     return image.extract_area(left, top, width, height);
   }
 
+  /*
+   * Calculate (a * in + b)
+   */
+  VImage Linear(VImage image, double const a, double const b) {
+    if (HasAlpha(image)) {
+      // Separate alpha channel
+      VImage imageWithoutAlpha = image.extract_band(0,
+        VImage::option()->set("n", image.bands() - 1));
+      VImage alpha = image[image.bands() - 1];
+      return imageWithoutAlpha.linear(a, b).bandjoin(alpha);
+    } else {
+      return image.linear(a, b);
+    }
+  }
 }  // namespace sharp